extract independent classes
