+++ src/content/operators/impls/EnableJumpMarkOperator.ts
@@ -1,0 +1,10 @@
+import Operator from "../Operator";
+import MarkKeyRepository from "../../repositories/MarkKeyRepository";
+
+export default class EnableJumpMarkOperator implements Operator {
+  constructor(private readonly repository: MarkKeyRepository) {}
+
+  async run(): Promise<void> {
+    this.repository.enableJumpMode();
+  }
+}
