+++ src/content/operators/impls/EnableSetMarkOperator.ts
@@ -1,0 +1,10 @@
+import Operator from "../Operator";
+import MarkKeyRepository from "../../repositories/MarkKeyRepository";
+
+export default class EnableSetMarkOperator implements Operator {
+  constructor(private readonly repository: MarkKeyRepository) {}
+
+  async run(): Promise<void> {
+    this.repository.enableSetMode();
+  }
+}
