+++ src/content/operators/impls/FindNextOperator.ts
@@ -1,0 +1,10 @@
+import Operator from "../Operator";
+import FindMasterClient from "../../client/FindMasterClient";
+
+export default class FindNextOperator implements Operator {
+  constructor(private readonly findMasterClient: FindMasterClient) {}
+
+  async run(): Promise<void> {
+    this.findMasterClient.findNext();
+  }
+}
