+++ src/content/operators/impls/MarkOperatorFactoryChain.ts
@@ -1,27 +1,13 @@
 import { inject, injectable } from "tsyringe";
+import EnableSetMarkOperator from "./EnableSetMarkOperator";
+import EnableJumpMarkOperator from "./EnableJumpMarkOperator";
 import Operator from "../Operator";
 import OperatorFactoryChain from "../OperatorFactoryChain";
 import MarkKeyRepository from "../../repositories/MarkKeyRepository";
 import * as operations from "../../../shared/operations";
 
-export class EnableSetMarkOperator implements Operator {
-  constructor(private readonly repository: MarkKeyRepository) {}
-
-  async run(): Promise<void> {
-    this.repository.enableSetMode();
-  }
-}
-
-export class EnableJumpMarkOperator implements Operator {
-  constructor(private readonly repository: MarkKeyRepository) {}
-
-  async run(): Promise<void> {
-    this.repository.enableJumpMode();
-  }
-}
-
 @injectable()
-export class MarkOperatorFactoryChain implements OperatorFactoryChain {
+export default class MarkOperatorFactoryChain implements OperatorFactoryChain {
   constructor(
     @inject("MarkKeyRepository")
     private readonly markKeyRepository: MarkKeyRepository
