+++ src/content/operators/impls/OperatorFactoryImpl.ts
@@ -1,17 +1,17 @@
 import { inject, injectable } from "tsyringe";
 import OperatorFactory from "../OperatorFactory";
-import { AddonOperatorFactoryChain } from "./addons";
-import { ClipboardOperatorFactoryChain } from "./clipboard";
-import { FindOperatorFactoryChain } from "./find";
-import { FocusOperatorFactoryChain } from "./focus";
-import { MarkOperatorFactoryChain } from "./mark";
-import { ScrollOperatorFactoryChain } from "./scroll";
-import { FollowOperatorFactoryChain } from "./follow";
 import BackgroundOperationOperator from "./BackgroundOperationOperator";
 import Operator from "../Operator";
 import OperatorFactoryChain from "../OperatorFactoryChain";
 import { Operation } from "../../../shared/operations";
 import OperationClient from "../../client/OperationClient";
+import AddonOperatorFactoryChain from "./AddonOperatorFactoryChain";
+import ClipboardOperatorFactoryChain from "./ClipboardOperatorFactoryChain";
+import FindOperatorFactoryChain from "./FindOperatorFactoryChain";
+import FocusOperatorFactoryChain from "./FocusOperatorFactoryChain";
+import FollowOperatorFactoryChain from "./FollowOperatorFactoryChain";
+import MarkOperatorFactoryChain from "./MarkOperatorFactoryChain";
+import ScrollOperatorFactoryChain from "./ScrollOperatorFactoryChain";
 
 @injectable()
 export default class OperatorFactoryImpl implements OperatorFactory {
