+++ src/content/operators/impls/ScrollToEndOperator.ts
@@ -1,0 +1,20 @@
+import AbstractScrollOperator from "./AbstractScrollOperator";
+import Operator from "../Operator";
+import ScrollPresenter from "../../presenters/ScrollPresenter";
+import SettingRepository from "../../repositories/SettingRepository";
+
+export default class ScrollToEndOperator
+  extends AbstractScrollOperator
+  implements Operator {
+  constructor(
+    private readonly presenter: ScrollPresenter,
+    settingRepository: SettingRepository
+  ) {
+    super(settingRepository);
+  }
+
+  async run(): Promise<void> {
+    const smooth = this.getSmoothScroll();
+    this.presenter.scrollToEnd(smooth);
+  }
+}
