+++ src/content/operators/impls/YankURLOperator.ts
@@ -1,0 +1,16 @@
+import Operator from "../Operator";
+import ClipboardRepository from "../../repositories/ClipboardRepository";
+import ConsoleClient from "../../client/ConsoleClient";
+
+export default class YankURLOperator implements Operator {
+  constructor(
+    private readonly repository: ClipboardRepository,
+    private readonly consoleClient: ConsoleClient
+  ) {}
+
+  async run(): Promise<void> {
+    const url = window.location.href;
+    this.repository.write(url);
+    await this.consoleClient.info("Yanked " + url);
+  }
+}
